allow multi threaded access to db and stmt (#34)
These objects are just references to their C pointers, thus
can be frozen safely. Multithreading on the SQLite is more
complicated and we might handle it once ksqlite3 has multi
threaded APIs but for now, it is best the core to allow
multi-threaded access and let upper layers manage it.